+++ apps/generic-issuance-client/src/pages/404.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
-import { PageContent, PodLink } from "../../components/Core";
-import { GlobalPageHeader } from "../../components/header/GlobalPageHeader";
+import { PageContent, PodLink } from "../components/Core";
+import { GlobalPageHeader } from "../components/header/GlobalPageHeader";
 
 export function NotFound(): ReactNode {
   return (
