+++ apps/generic-issuance-client/src/pages/pipeline/DetailsSections/PipelineLatestConsumersSection.tsx
