+++ apps/generic-issuance-client/src/pages/pipeline/DetailsSections/PipelineManualTicketsSection.tsx
@@ -1,0 +1,22 @@
+import {
+  PipelineDefinition,
+  PipelineInfoResponseValue
+} from "@pcd/passport-interface";
+import { ReactNode } from "react";
+
+export function PipelineManualTicketsSection({
+  pipeline
+}: {
+  pipelineInfo: PipelineInfoResponseValue;
+  pipeline: PipelineDefinition;
+  isAdminView: boolean;
+}): ReactNode {
+  return (
+    <div>
+      manual tickets section
+      <br />
+      <br />
+      {pipeline.id}
+    </div>
+  );
+}
